requirements.txt
streamlit>=1.25.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
librosa>=0.10.0
joblib>=1.3.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.8.0